requirements.txt
autopep8==1.4.4
certifi==2019.9.11
chardet==3.0.4
Click==7.0
cycler==0.10.0
fdb==2.0.1
Flask==1.1.1
Flask-SocketIO==4.2.1
future==0.18.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.0.3
mysql-connector-python==8.0.18
nltk==3.4.5
numpy==1.17.3
oauthlib==3.1.0
pandas==0.25.2
protobuf==3.10.0
pycodestyle==2.5.0
pyodbc==4.0.27
pyparsing==2.4.2
PySocks==1.7.1
python-dateutil==2.8.0
python-engineio==3.10.0
python-socketio==4.3.1
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
textblob==0.15.3
tweepy==3.8.0
urllib3==1.25.6
Werkzeug==0.16.0
